Superseded bash scrit with single python for hyperdex and main benchmarks.
Added retrieving range entry-by-entry analysis data with a plot.

+++ scripts/main_benchmark.py
@@ -29,14 +29,23 @@ env.warn_only = True
 
 DTACH_COMMAND = 'dtach -n `mktemp -u /tmp/dtach.XXXX`'
 
-if len(sys.argv) < 4:
-    print "Please specify the following args: chunk_size, repetitions_num, aws_access_key_id, aws_secret_access_key"
+if len(sys.argv) < 5:
+    print "Please specify args"
     sys.exit(0)
 
+EXPERIMENT_TYPE = sys.argv[1]
+CHUNK_SIZE = sys.argv[2]
+REPETITIONS_NUM = sys.argv[3]
+AWS_ACCESS_KEY_ID = sys.argv[4]
+AWS_SECRET_ACCESS_KEY = sys.argv[5]
+TEST_RANGE = sys.argv[4]
+TEST_RANGE_ENTRY = sys.argv[5]
+
 @parallel
-def wipe_state():
-    command = 'rm -rf {}/*'.format(TEMP_DIR)
-    run(command)
+def wipe_state(remove_logs):
+    if remove_logs:
+        command = 'rm -rf {}/*'.format(TEMP_DIR)
+        run(command)
 
     command = 'pkill -f hyperdex'
     run(command)
@@ -44,17 +53,28 @@ def wipe_state():
     command = 'sudo service cassandra stop'
     run(command)
 
-@parallel
 def start_coordinator():
-    command = '{} hyperdex coordinator -f -p 1982 -D {}/hyperdex_data'.format(DTACH_COMMAND, TEMP_DIR)
+    if EXPERIMENT_TYPE == 'Main':
+        command = '{} hyperdex coordinator -f -p 1982 -D {}/hyperdex_data'.format(DTACH_COMMAND, TEMP_DIR)
+    elif EXPERIMENT_TYPE == 'HyperDex':
+        command = '{} hyperdex coordinator -f -l 127.0.0.1 -p 1982 -D {}/hyperdex_data'.format(DTACH_COMMAND, TEMP_DIR)
     run(command)
 
-def start_daemon():
-    command = '{} hyperdex daemon -f --listen-port=2014 --coordinator={} --coordinator-port=1982 --data={}/hyperdex_daemon'.format(DTACH_COMMAND, COORDINATOR_IP, TEMP_DIR)
+@parallel
+def start_daemon(daemon_port):
+    if EXPERIMENT_TYPE == 'Main':
+        coordinator_ip = COORDINATOR_IP
+    elif EXPERIMENT_TYPE == 'HyperDex':
+        coordinator_ip = '127.0.0.1'
+    command = '{} hyperdex daemon -f --listen-port={} --coordinator={} --coordinator-port=1982 --data={}/hyperdex_daemon{}'.format(DTACH_COMMAND, daemon_port, coordinator_ip, TEMP_DIR, daemon_port)
     run(command)
 
 def setup_spaces():
-    command = 'python ~/solitude/scripts/setup_spaces.py {}'.format(COORDINATOR_IP)
+    if EXPERIMENT_TYPE == 'Main':
+        ip = COORDINATOR_IP
+    elif EXPERIMENT_TYPE == 'HyperDex':
+        ip = '127.0.0.1'
+    command = 'python ~/solitude/scripts/setup_spaces.py {}'.format(ip)
     run(command)
 
 @parallel
@@ -62,12 +82,16 @@ def start_cassandra():
     command = 'sudo service cassandra start'
     run (command)
 
-def start_experiment(chunk_size, repetitions_num, aws_access_key_id, aws_secret_access_key):
+def start_experiment():
+
     class_path = '.:/usr/local/share/java/org.hyperdex.client-1.8.1.jar:lib/aws-sdk/lib/aws-java-sdk-1.11.255.jar:lib/aws-sdk/third-party/lib/*:lib/cassandra/cassandra-driver-core-3.3.2.jar:lib/cassandra/lib/*'
     command = 'cd ~/solitude && javac -classpath "{}" $(find . -name \'*.java\')'.format(class_path)
     run(command)
 
-    command = 'cd ~/solitude && java -classpath "{}" -Djava.library.path=/usr/local/lib ch/michel/test/MainBenchmark {} {} {} {} {} {}'.format(class_path, chunk_size, repetitions_num, COORDINATOR_IP, aws_access_key_id, aws_secret_access_key, COORDINATOR_IP)
+    if EXPERIMENT_TYPE == 'Main':
+        command = 'cd ~/solitude && java -classpath "{}" -Djava.library.path=/usr/local/lib ch/michel/test/MainBenchmark {} {} {} {} {} {}'.format(class_path, CHUNK_SIZE, REPETITIONS_NUM, COORDINATOR_IP, AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, COORDINATOR_IP)
+    elif EXPERIMENT_TYPE == 'HyperDex':
+        command = 'cd ~/solitude && java -classpath "{}" -Djava.library.path=/usr/local/lib ch/michel/test/HyperDexBenchmark {} {} {} {} >> temp/hyperdex_benchmark.log'.format(class_path, CHUNK_SIZE, REPETITIONS_NUM, TEST_RANGE, TEST_RANGE_ENTRY)
     run(command)
 
 # 0. Wipe all ex-state
@@ -76,7 +100,7 @@ def start_experiment(chunk_size, repetitions_num, aws_access_key_id, aws_secret_
 # 3. Connect to machine that will run an experiment, run java
 
 print green('Wiping any previous execution state')
-execute(wipe_state, hosts=ALL_HOSTS)
+execute(wipe_state, True, hosts=ALL_HOSTS)
 
 print green('Starting HyperDex coordinator')
 execute(start_coordinator, hosts=[COORDINATOR])
@@ -85,17 +109,26 @@ print green('Setting HyperDex up spaces')
 execute(setup_spaces, hosts=[COORDINATOR])
 
 print green('Starting HyperDex daemons')
-execute(start_daemon, hosts=DAEMONS)
+if EXPERIMENT_TYPE == 'Main':
+    execute(start_daemon, 2014, hosts=DAEMONS)
+elif EXPERIMENT_TYPE == 'HyperDex':
+    for port in range(2014, 2018):
+        execute(start_daemon, port, hosts=[COORDINATOR])
 
-print green('Starting Cassandra seed nodes')
-execute(start_cassandra, hosts=[COORDINATOR])
+if EXPERIMENT_TYPE == 'Main':
+    print green('Starting Cassandra seed nodes')
+    execute(start_cassandra, hosts=[COORDINATOR])
 
-print green('Starting Cassandra non-seed nodes')
-execute(start_cassandra, hosts=DAEMONS)
+    print green('Starting Cassandra non-seed nodes')
+    execute(start_cassandra, hosts=DAEMONS)
 
 time.sleep(60) # Give time for the datastores to spin up
+
 print green('Running experiment')
-execute(start_experiment, sys.argv[1], sys.argv[2], sys.argv[3], sys.argv[4], hosts=[EXPERIMENT])
+if EXPERIMENT_TYPE == 'Main':
+    execute(start_experiment, hosts=[EXPERIMENT])
+elif EXPERIMENT_TYPE == 'HyperDex':
+    execute(start_experiment, hosts=[COORDINATOR])
 
 print green('Wipe the state')
-execute(wipe_state, hosts=ALL_HOSTS)
+execute(wipe_state, False, hosts=ALL_HOSTS)
